requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crypto==1.4.1
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Naked==0.1.32
node==1.2.2
odict==1.9.0
packaging==24.2
pluggy==1.5.0
plumber==1.7
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
PyNaCl==1.5.0
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.2
shellescape==3.8.1
soupsieve==2.6
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.1
zope.event==5.0
zope.hookable==7.0
zope.interface==7.2
zope.lifecycleevent==5.1
zope.proxy==6.1
